[back] refactor: gather tests related to the accounts API
it should be easier to find them and maintain them

each APIs' TestCase are saved in their own file, and
the accounts API tests didn't follow this rule for no
particular reason
