+++ backend/core/tests/test_api_accounts.py
@@ -1,3 +1,5 @@
+from unittest.mock import patch
+
 from django.test import TestCase
 from rest_framework import status
 from rest_framework.test import APIClient
@@ -13,6 +15,7 @@ class AccountsRegisterTestCase(TestCase):
     Even if this API is provided by a third-party package, its default
     behaviour has been customized, and thus needs to be tested.
     """
+
     _existing_username = "existing"
     _existing_email = "existing@example.org"
     _existing_email_alt = "Existing@example.org"
@@ -119,3 +122,32 @@ class AccountsRegisterTestCase(TestCase):
             format="json",
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+
+class AccountsResetPasswordTestCase(TestCase):
+    """
+    TestCase of the /accounts/reset-password/ API.
+
+    Even if this API is provided by a third-party package, its default
+    behaviour has been customized, and thus needs to be tested.
+    """
+
+    @patch("rest_registration.utils.signers.DataSigner.verify", new=lambda x: True)
+    def test_reset_password_force_user_activation(self):
+        client = APIClient()
+        user = UserFactory(is_active=False)
+
+        resp = client.post(
+            "/accounts/reset-password/",
+            data={
+                "user_id": user.id,
+                "password": "very_uncommon_pwd",
+                "timestamp": 123,
+                "signature": "abc",
+            },
+            format="json",
+        )
+        self.assertEqual(resp.status_code, 200, resp.content)
+
+        user.refresh_from_db()
+        self.assertTrue(user.is_active)
